CInterface: Added an example for testing the C interface.
pi_sender.c creates a broker and a value federate. The federate publishes t*pi
every 0.01 sec.	    for 20 time-steps.
./pi_sender

pi_receiver creates a value federate to receive the values published by pi_sender.
./pi_receiver

There are a few	things that need to be sorted out for this example to be fully ready:
i) When launching the pi_receiver, sometimes it gets stuck with the message 'Address already in use'.
ii) The	first and last values received by pi_receiver are incorrect. The receiver never receives the value at t=0 and has a copy of the previous time-step value for t=0.19.
iii) The pi_receiver federate gets stuck in helicsFinalize() so the program hangs after sending/receiving of messages.