Add .git directory detection to fix fatal error
When no git directory is configured, cmake would still
attempt to call the logic disabling git hooks. This
resulted in a fatal error for the user.
To resolve this, a check for the present of the .git
directory in the cmake project souce directory was
added, and a status message that hooks are inactive
when not in a git repository was added.