Normalize source_only
Standardized terms: source_only, sourceonly, sourceOnly

Deprecated terms: source

Given that "source" is a VERY common term in the code base, I'm not certain that I've caught all the places it should no longer be used. The good news is that the tests seem to prefer the term "source_only" so there's a chance there wasn't much to change.