Catch-up on normalizations
A few overlooked normalizations and typos are corrected.

This commit doesn't pass all the built-in tests but it doesn't seem like this is related to the normalization work (unable to find configuration files). This commit will test the CI build to see if its just something on my system (Phil doesn't see these errors).