Adding HELICS theming to install diagram
Still can't get rid of the crazy margins, though. If you set "useMaxWidth" to "false" the diagram fills the whitespace but bleeds off the edge of the page. Its clear that the space reserved on the page is for a diagram this size but the rendered image ends up being smaller.